requirements.txt
streamlit ==1.41.1
python-dotenv ==1.0.1
